+++ leda/components/Input/handlers.ts
@@ -1,23 +1,25 @@
 import * as React from 'react';
 
+import { isFunction, isNil } from 'lodash';
 import { CustomEventHandler, SetState } from '../../commonTypes';
-import { InputProps } from './types';
+import { AdjustCursor, InputProps } from './types';
 import {
-  isSymbolAllowed, isSymbolForbidden, transformToCase,
+  isSymbolAllowed, isSymbolForbidden, transformToCase, getSelection, getNewValueLength, getMaxPastedLength, getNewPastedValue,
 } from './helpers';
+import { stringToMaxLength } from '../../utils';
 
 /**
  * Function creates change event handler
  * @param {InputProps} props - properties of component
  * @param {SetState<string>} setValue - set state action
- * @param {(event: React.ChangeEvent<HTMLInputElement> | React.ClipboardEvent<HTMLInputElement>) => void} setCursorToPrevPosition - function returns cursor to initial position
+ * @param {AdjustCursor} adjustCursor - function sets cursor to specified or previous position
  *
  * @returns {CustomEventHandler<React.ChangeEvent<HTMLInputElement>>} - change event handler
  */
 export const createChangeHandler = (
   props: InputProps,
   setValue: SetState<string>,
-  setCursorToPrevPosition: (event: React.ChangeEvent<HTMLInputElement> | React.ClipboardEvent<HTMLInputElement>) => void,
+  adjustCursor: AdjustCursor,
 ): CustomEventHandler<React.ChangeEvent<HTMLInputElement>> => (event) => {
   const { value } = event.target;
   const {
@@ -30,12 +32,14 @@ export const createChangeHandler = (
 
   // Next block prevents from replacing previous value if maxLength is exceeded
   // and sets cursor back to initial position
-  if (maxLength != null && value.length > maxLength) {
-    setCursorToPrevPosition(event);
+  if (!isNil(maxLength) && value.length > maxLength) {
+    adjustCursor(event);
     return;
   }
 
-  const newValue = letterCase ? transformToCase(value, letterCase) : value;
+  const maxLengthAdjustedValue = stringToMaxLength(value, maxLength);
+
+  const newValue = letterCase ? transformToCase(value, letterCase) : maxLengthAdjustedValue;
 
   if (props.value === undefined) {
     setValue(newValue);
@@ -176,26 +180,71 @@ export const createResetHandler = (
 };
 
 /**
- * Function creates paste event handler which checks if new value exceeds maxLength
- * and in this case returns cursor to initial position.
- * Paste handler is needed, because input.selectionStart value is counted differently depending on event type
+ * Function creates paste event handler which takes control if new value exceeds maxLength
  *
  * @param {InputProps} props - properties of component
- * @param {(event: React.ChangeEvent<HTMLInputElement> | React.ClipboardEvent<HTMLInputElement>) => void} setCursorToPrevPosition - function sets cursor to previous position if maxLength is exceeded
+ * @param {string} value - current value of input
+ * @param {SetState<string>} setValue - set state function for input value
+ * @param {AdjustCursor} adjustCursor - function sets cursor to specified or previous position if maxLength is exceeded
  *
  * @returns {CustomEventHandler<React.ClipboardEvent<HTMLInputElement>>} - paste event handler
  */
 export const createPasteHandler = (
   props: InputProps,
-  setCursorToPrevPosition: (event: React.ChangeEvent<HTMLInputElement> | React.ClipboardEvent<HTMLInputElement>) => void,
+  value: string,
+  setValue: SetState<string>,
+  adjustCursor: AdjustCursor,
 ): CustomEventHandler<React.ClipboardEvent<HTMLInputElement>> => (event) => {
-  const { value } = event.target as HTMLInputElement;
-  const { maxLength } = props;
+  const {
+    maxLength, letterCase, isDisabled, onChange, name,
+  } = props;
+  const inputElement = event.target as HTMLInputElement;
 
+  if (isDisabled) return;
+
+  const { selectionStart, selectionEnd, selectedRange } = getSelection(inputElement);
   const pastedValue = event.clipboardData.getData('Text');
-  const newValueLength = value.length + pastedValue.length;
+  const newValueLength = getNewValueLength(value, pastedValue, selectedRange);
 
-  if (maxLength != null && newValueLength > maxLength) {
-    setCursorToPrevPosition(event);
+  if (!isNil(maxLength) && newValueLength > maxLength) {
+    const maxPastedLength = getMaxPastedLength(value, maxLength, selectedRange);
+
+    const adjustedPastedValue = stringToMaxLength(pastedValue, maxPastedLength);
+
+    const newValue = getNewPastedValue({
+      adjustedPastedValue,
+      maxLength,
+      oldValue: value,
+      selectedRange,
+      selectionEnd,
+      selectionStart,
+    });
+
+    if (newValue !== null) {
+      event.preventDefault();
+
+      const adjustedNewValue = letterCase ? transformToCase(newValue, letterCase) : newValue;
+
+      if (props.value === undefined) {
+        setValue(adjustedNewValue);
+      }
+
+      if (isFunction(onChange)) {
+        const customEvent = {
+          ...event,
+          component: {
+            name,
+            value: adjustedNewValue,
+          },
+          target: event.target as HTMLInputElement,
+        };
+
+        onChange(customEvent);
+      }
+
+      if (selectionStart !== null) adjustCursor(event, selectionStart + maxPastedLength);
+    } else {
+      adjustCursor(event);
+    }
   }
 };