+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Completed description in README.md: technologies, git workflow, project's structure
+### Fixed
+- Input: Fixed paste event handler for Input with maxLength prop
 
 
 ## [0.35.0] - 2020-11-23